fix(tui): split mount.tsx from tuiLogger so opentui never loads on Node
@opentui/core statically imports .scm tree-sitter query files, which Node's
ESM loader rejects with "Unknown file extension". Even though the renderer
itself was already behind a dynamic import, App.tsx pulled in opentui's JSX
runtime via jsxImportSource at the top of tuiLogger.tsx's static graph,
crashing import('@kubb/tui') the moment --tui was passed under Node.

- tuiLogger is now a plain .ts file with no JSX and no static references to
  opentui. It runs the Bun/TTY guard first, then dynamically imports the
  separate mount.tsx module which holds App, createCliRenderer, and createRoot.
- The build now emits mount as its own chunk, so non-Bun runtimes can load
  tuiLogger, hit the guard, and bail without ever touching opentui's import
  tree.
- Failures inside mount() (createCliRenderer, createRoot) are surfaced on
  stderr with the underlying message instead of being swallowed.
