+++ packages/tui/src/index.ts
@@ -1,3 +1,3 @@
-export { tuiLogger } from './tuiLogger.tsx'
+export { tuiLogger } from './tuiLogger.ts'
 export { getTuiUnavailableReason, isBun, isInteractiveTTY } from './runtime.ts'
 export type { TuiAction, TuiState, PluginEntry, PluginStatus, LogEntry, LogLevel, HookEntry, RunStatus } from './state.ts'
