+++ packages/tui/src/tuiLogger.ts
@@ -1,15 +1,11 @@
-import process from 'node:process'
-import React from 'react'
 import { defineLogger, type LoggerOptions, logLevel as logLevelMap } from '@kubb/core'
-import { App } from './App.tsx'
 import { getTuiUnavailableReason } from './runtime.ts'
-import type { TuiAction } from './state.ts'
+import type { Mount } from './mount.tsx'
 
 /**
- * Output sink for a hook subprocess. Mirrors the contract that
- * `@kubb/cli`'s `HookSinkFactory` already uses with the clack logger so the
- * generation runner can route stdout/stderr without knowing which logger is
- * mounted.
+ * Output sink for a hook subprocess. Mirrors the contract `@kubb/cli`'s
+ * `HookSinkFactory` already uses with the clack logger so the generation
+ * runner can route stdout/stderr without knowing which logger is mounted.
  */
 type HookSinkOptions = {
   stream?: boolean
@@ -19,34 +15,6 @@ type HookSinkOptions = {
 }
 
 type HookSinkFactory = (commandWithArgs: string, hookId: string) => HookSinkOptions | null
-
-type ReactRoot = {
-  render: (node: React.ReactNode) => void
-  unmount: () => void
-}
-
-type CliRenderer = {
-  start: () => void
-  stop: () => void
-  destroy: () => void
-}
-
-type OpenTuiCore = {
-  createCliRenderer: (config?: { exitOnCtrlC?: boolean; targetFps?: number; useMouse?: boolean }) => Promise<CliRenderer>
-}
-
-type OpenTuiReact = {
-  createRoot: (renderer: CliRenderer) => ReactRoot
-}
-
-async function loadOpenTui(): Promise<{ core: OpenTuiCore; react: OpenTuiReact } | null> {
-  try {
-    const [core, react] = await Promise.all([import('@opentui/core'), import('@opentui/react')])
-    return { core: core as unknown as OpenTuiCore, react: react as unknown as OpenTuiReact }
-  } catch {
-    return null
-  }
-}
 
 /**
  * Pulls the plugin list from `config` without depending on `@kubb/core`'s
@@ -63,75 +31,41 @@ function getPluginNames(config: unknown): Array<string> {
  * Terminal UI logger backed by opentui + React. Activates only when the
  * runtime is Bun and stdout is a TTY; otherwise `install` returns `null` and
  * the caller is expected to fall back to another logger (clack/plain).
+ *
+ * This module deliberately avoids any static import of opentui or the React
+ * tree. On Node, opentui's `.scm` tree-sitter assets can't be loaded by the
+ * ESM loader, so we keep the JSX and renderer behind a dynamic import that
+ * runs only after the Bun guard.
  */
 export const tuiLogger = defineLogger<LoggerOptions, HookSinkFactory | null>({
   name: 'tui',
   async install(context, options) {
     const reason = getTuiUnavailableReason()
     if (reason) {
-      // No logger has subscribed yet when install runs, so route the message
-      // straight to stderr to make sure the user sees it before the fallback
-      // logger takes over.
       console.error(`[kubb] ${reason}`)
       return null
     }
 
-    const opentui = await loadOpenTui()
-    if (!opentui) {
-      console.error('[kubb] Could not load @opentui/core or @opentui/react. Install @kubb/tui peer dependencies or omit --tui.')
+    let mountFn: typeof import('./mount.tsx').mount
+    try {
+      ;({ mount: mountFn } = await import('./mount.tsx'))
+    } catch (error) {
+      const message = error instanceof Error ? error.message : String(error)
+      console.error(`[kubb] Failed to load the TUI renderer: ${message}`)
       return null
     }
 
-    const renderer = await opentui.core.createCliRenderer({ exitOnCtrlC: false, useMouse: false, targetFps: 60 })
-    const root = opentui.react.createRoot(renderer)
+    let mounted: Mount
+    try {
+      mounted = await mountFn()
+    } catch (error) {
+      const message = error instanceof Error ? error.message : String(error)
+      console.error(`[kubb] opentui failed to start: ${message}`)
+      return null
+    }
 
+    const { dispatch, teardown } = mounted
     const logLevel = options?.logLevel ?? logLevelMap.info
-    const listeners = new Set<(action: TuiAction) => void>()
-
-    function dispatch(action: TuiAction): void {
-      for (const listener of listeners) listener(action)
-    }
-
-    function subscribe(listener: (action: TuiAction) => void): () => void {
-      listeners.add(listener)
-      return () => {
-        listeners.delete(listener)
-      }
-    }
-
-    root.render(<App subscribe={subscribe} />)
-    renderer.start()
-
-    let torn = false
-    function teardown(): void {
-      if (torn) return
-      torn = true
-      try {
-        root.unmount()
-      } catch {
-        // ignore
-      }
-      try {
-        renderer.stop()
-      } catch {
-        // ignore
-      }
-      try {
-        renderer.destroy()
-      } catch {
-        // ignore
-      }
-    }
-
-    process.once('exit', teardown)
-    process.once('SIGINT', () => {
-      teardown()
-      process.exit(130)
-    })
-    process.once('SIGTERM', () => {
-      teardown()
-      process.exit(143)
-    })
 
     context.on('kubb:lifecycle:start', ({ version }) => {
       dispatch({ type: 'lifecycle:start', version })
